requirements.txt
ipykernel
pandas
numpy
matplotlib
plotly
seaborn
scipy
scikit-learn
imblearn
xgboost
catboost

pymysql
from_root
dvc
dill
PyYAML

python-multipart


cryptography
pymysql[crypto]

-e .